% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_aucs}
\alias{bootstrap_aucs}
\title{TO BE EDITED.}
\usage{
bootstrap_aucs(estimator, X, y, n_divisions = 1000, n_iterations = 100)
}
\arguments{
\item{estimator}{TO BE EDITED.}

\item{X}{TO BE EDITED.}

\item{y}{TO BE EDITED.}

\item{n_divisions}{TO BE EDITED.}

\item{n_iterations}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

