% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretability.R
\name{calculate_shap}
\alias{calculate_shap}
\title{Calcular Valores SHAP}
\usage{
calculate_shap(
  final_fit,
  train_data,
  test_data,
  target,
  n_samples = 100,
  verbose = TRUE
)
}
\description{
Calcular Valores SHAP
}
