% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arguments}
\alias{check_arguments}
\title{Check the positional arguments.}
\usage{
check_arguments(.data, dependent_variable, algorithm)
}
\arguments{
\item{.data}{A data.frame; the data to be analyzed.}

\item{dependent_variable}{A character vector of length one; the dependent variable for this analysis.}

\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}
}
\value{
NULL.
}
\description{
Check the positional arguments.
}
\seealso{
Other utils: \code{\link{correlation_test}},
  \code{\link{process_coefficients}},
  \code{\link{process_variable_importances}},
  \code{\link{reduce_cores}},
  \code{\link{remove_variables}}
}

