% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cocaine}
\alias{cocaine}
\title{Cocaine data.}
\format{Data frame with columns
\describe{
\item{subject}{To Be Edited}
\item{DIAGNOSIS}{To Be Edited}
\item{AGE}{To Be Edited}
\item{Male}{To Be Edited}
\item{EDU_YRS}{To Be Edited}
\item{IMT_COMM_errors}{To Be Edited}
\item{IMT_OMIS_errors}{To Be Edited}
\item{A_IMT}{To Be Edited}
\item{B_D_IMT}{To Be Edited}
\item{Stop_ssrt}{To Be Edited}
\item{lnk_adjdd}{To Be Edited}
\item{lkhat_Kirby}{To Be Edited}
\item{REVLR_per_errors}{To Be Edited}
\item{BIS_attention}{To Be Edited}
\item{BIS_motor}{To Be Edited}
\item{BIS_NonPL}{To Be Edited}
\item{IGT_TOTAL}{To Be Edited}
}}
\source{
https://www.tobeedited.com/
}
\usage{
cocaine
}
\description{
Cocaine data.
}
\examples{
  cocaine
}
\keyword{datasets}

