% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cocaine_dependence}
\alias{cocaine_dependence}
\title{Cocaine data.}
\format{Data frame with columns
\describe{
\item{subject}{To Be Edited}
\item{diagnosis}{To Be Edited}
\item{age}{To Be Edited}
\item{male}{To Be Edited}
\item{edu_yrs}{To Be Edited}
\item{imt_comm_errors}{To Be Edited}
\item{imt_omis_errors}{To Be Edited}
\item{a_imt}{To Be Edited}
\item{b_d_imt}{To Be Edited}
\item{stop_ssrt}{To Be Edited}
\item{lnk_adjdd}{To Be Edited}
\item{lkhat_kirby}{To Be Edited}
\item{revlr_per_errors}{To Be Edited}
\item{bis_attention}{To Be Edited}
\item{bis_motor}{To Be Edited}
\item{bis_nonpL}{To Be Edited}
\item{igt_total}{To Be Edited}
}}
\source{
https://www.tobeedited.com/
}
\usage{
cocaine_dependence
}
\description{
Cocaine data.
}
\examples{
  cocaine_dependence
}
\keyword{datasets}

