% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{compare_models}
\alias{compare_models}
\title{Comparar Resultados de Modelos}
\usage{
compare_models(cv_results, task, select_metric = NULL, verbose = TRUE)
}
\description{
Comparar Resultados de Modelos
}
