% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mtest.R
\name{cor_mtest}
\alias{cor_mtest}
\title{TO BE EDITED.}
\usage{
cor_mtest(mat, conf_level = 0.95)
}
\arguments{
\item{mat}{TO BE EDITED.}

\item{conf_level}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

