% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correlation_test}
\alias{correlation_test}
\title{Compute the matrix of p-value.}
\usage{
correlation_test(x, confidence_level = 0.95, ...)
}
\arguments{
\item{x}{An object of class "DataFrame" or "matrix".}

\item{confidence_level}{confidence level for the returned confidence interval. Currently only used for the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.}

\item{...}{further arguments to be passed to or from \code{cor.test}.}
}
\value{
A list containing three matrices; p_value, lower_bound, and upper bound.
}
\description{
See here for source: \url{http://www.sthda.com/english/wiki/visualize-correlation-matrix-using-correlogram}.
}
\seealso{
Other utils: \code{\link{process_coefficients}},
  \code{\link{process_variable_importances}},
  \code{\link{reduce_cores}},
  \code{\link{remove_variables}}
}
