% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{easy_glmnet}
\alias{easy_glmnet}
\title{TO BE EDITED.}
\usage{
easy_glmnet(.data, dependent_variable = NULL, family = "gaussian",
  exclude_variables = NULL, train_size = 0.667, n_divisions = 1000,
  n_iterations = 10, n_samples = 1000, out_directory = ".",
  random_state = NULL, ...)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

