% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{easy_random_forest}
\alias{easy_random_forest}
\title{TO BE EDITED.}
\usage{
easy_random_forest(.data, dependent_variable, family = "gaussian",
  sampler = NULL, preprocessor = NULL, exclude_variables = NULL,
  categorical_variables = NULL, train_size = 0.667,
  survival_rate_cutoff = 0.05, n_samples = 1000, n_divisions = 1000,
  n_iterations = 10, out_directory = ".", random_state = NULL,
  progress_bar = TRUE, n_core = 1, ...)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

