% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{easy_support_vector_machine}
\alias{easy_support_vector_machine}
\title{TO BE EDITED.}
\usage{
easy_support_vector_machine(.data, dependent_variable, family = "gaussian",
  resample = NULL, preprocess = NULL, measure = NULL,
  exclude_variables = NULL, categorical_variables = NULL,
  train_size = 0.667, survival_rate_cutoff = 0.05, n_samples = 1000,
  n_divisions = 1000, n_iterations = 10, random_state = NULL,
  progress_bar = TRUE, n_core = 1, ...)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

