% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyGlmnet.R
\docType{package}
\name{easyGlmnet}
\alias{easyGlmnet}
\alias{easyGlmnet-package}
\title{easyGlmnet: Easily build and evaluate penalized regression models.}
\description{
easyGlmnet: Easily build and evaluate penalized regression models.
}

