% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyML.R
\docType{package}
\name{easyML}
\alias{easyML}
\alias{easyML-package}
\title{easyML: Easily build and evaluate penalized regression models.}
\description{
easyML: Easily build and evaluate penalized regression models.
}

