% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_categorical_basic}
\alias{eda_categorical_basic}
\title{Analisis de variables categoricas (basico)}
\usage{
eda_categorical_basic(data, verbose = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{verbose}{Mostrar resultados}
}
\value{
Lista con informacion de variables categoricas
}
\description{
Analisis de variables categoricas (basico)
}
