% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_correlation_basic}
\alias{eda_correlation_basic}
\title{Matriz de correlacion (basico)}
\usage{
eda_correlation_basic(data, verbose = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{verbose}{Mostrar resultados}
}
\value{
Matriz de correlacion
}
\description{
Matriz de correlacion (basico)
}
