% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_numeric_basic}
\alias{eda_numeric_basic}
\title{Analisis de variables numericas (basico)}
\usage{
eda_numeric_basic(data, verbose = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{verbose}{Mostrar resultados}
}
\value{
Data frame con estadisticos
}
\description{
Analisis de variables numericas (basico)
}
