% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_outliers}
\alias{eda_outliers}
\title{Deteccion de outliers}
\usage{
eda_outliers(data, verbose = TRUE, iqr_multiplier = 3)
}
\arguments{
\item{data}{Data frame}

\item{verbose}{Mostrar resultados}

\item{iqr_multiplier}{Multiplicador del IQR (default: 3 para ser conservador)}
}
\description{
Deteccion de outliers
}
