% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_structure}
\alias{eda_structure}
\title{Estructura de los datos}
\usage{
eda_structure(data, verbose = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{verbose}{Mostrar resultados}
}
\value{
Lista con informacion estructural
}
\description{
Estructura de los datos
}
