% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_target_classification}
\alias{eda_target_classification}
\title{Distribucion del target (clasificacion)}
\usage{
eda_target_classification(data, target, verbose = TRUE)
}
\description{
Distribucion del target (clasificacion)
}
