% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{extract_coefficients}
\alias{extract_coefficients}
\title{Extract coefficients.}
\usage{
extract_coefficients(object)
}
\arguments{
\item{object}{A list of class \code{easy_*}, where * is the name of the algorithm.}
}
\value{
A data.frame, the replicated coefficients.
}
\description{
Extract coefficients.
}

