% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{extract_variable_importances.easy_random_forest}
\alias{extract_variable_importances.easy_random_forest}
\title{Extract variable importance scores from a random forest model.}
\usage{
\method{extract_variable_importances}{easy_random_forest}(object)
}
\arguments{
\item{object}{A list of class \code{easy_random_forest}.}
}
\value{
A data.frame, the replicated random forest variable importance scores.
}
\description{
This function wraps the procedure for extracting variable importances
 from a random forest model and makes it accessible 
to the easyml core framework.
}
