% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{extract_variable_importances}
\alias{extract_variable_importances}
\title{Extract variable importances.}
\usage{
extract_variable_importances(object)
}
\arguments{
\item{object}{A list of class \code{easy_*}, where * is the name of the algorithm.}
}
\value{
A data.frame, the replicated random forest variable importance scores.
}
\description{
Extract variable importances.
}

