% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avNNet.R
\name{fit_model.easy_avNNet}
\alias{fit_model.easy_avNNet}
\title{Fit a neural network model.}
\usage{
\method{fit_model}{easy_avNNet}(object)
}
\arguments{
\item{object}{A list of class \code{easy_avNNet}.}
}
\value{
A list of class \code{easy_avNNet}.
}
\description{
Fit a neural network model.
}
