% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_neural_network.R
\name{fit_model.easy_deep_neural_network}
\alias{fit_model.easy_deep_neural_network}
\title{Fit a deep neural network model.}
\usage{
\method{fit_model}{easy_deep_neural_network}(object)
}
\arguments{
\item{object}{A list of class \code{easy_deep_neural_network}.}
}
\value{
A list of class \code{easy_deep_neural_network}.
}
\description{
This function wraps the procedure for fitting a 
deep neural network model and makes it accessible 
to the easyml core framework.
}
