% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.R
\name{fit_model.easy_neural_network}
\alias{fit_model.easy_neural_network}
\title{Fit a neural network model.}
\usage{
\method{fit_model}{easy_neural_network}(object)
}
\arguments{
\item{object}{A list of class \code{easy_neural_network}.}
}
\value{
A list of class \code{easy_neural_network}.
}
\description{
Fit a neural network model.
}

