% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{fit_model.easy_random_forest}
\alias{fit_model.easy_random_forest}
\title{Fit a random forest model.}
\usage{
\method{fit_model}{easy_random_forest}(object)
}
\arguments{
\item{object}{A list of class \code{easy_random_forest}.}
}
\value{
A list of class \code{easy_random_forest}.
}
\description{
This function wraps the procedure for fitting a 
random forest model and makes it accessible 
to the easyml core framework.
}
