% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{fit_model.easy_support_vector_machine}
\alias{fit_model.easy_support_vector_machine}
\title{Fit a support vector machine regression model.}
\usage{
\method{fit_model}{easy_support_vector_machine}(object)
}
\arguments{
\item{object}{A list of class \code{easy_support_vector_machine}.}
}
\value{
A list of class \code{easy_support_vector_machine}.
}
\description{
This function wraps the procedure for fitting a 
support vector machine model and makes it accessible 
to the easyml core framework.
}
