% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fit_model}
\alias{fit_model}
\title{Fit model.}
\usage{
fit_model(X, y, ...)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format (inherit from class "sparseMatrix" as in package Matrix; not yet available for family="cox")}

\item{y}{response variable. Quantitative for family="gaussian", or family="poisson" (non-negative counts). For family="binomial" should be either a factor with two levels, or a two-column matrix of counts or proportions (the second column is treated as the target class; for a factor, the last level in alphabetical order is the target class). For family="multinomial", can be a nc>=2 level factor, or a matrix with nc columns of counts or proportions. For either "binomial" or "multinomial", if y is presented as a vector, it will be coerced into a factor. For family="cox", y should be a two-column matrix with columns named 'time' and 'status'. The latter is a binary variable, with '1' indicating death, and '0' indicating right censored. The function Surv() in package survival produces such a matrix. For family="mgaussian", y is a matrix of quantitative responses.}

\item{...}{Arguments to be passed to \code{\link[glmnet]{glmnet}}. See that function's documentation for more details.}
}
\value{
A list, the model and the cross validated model.
}
\description{
Fit model.
}

