% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_easyml.R
\name{generate_all_plots}
\alias{generate_all_plots}
\title{Generar Todos los Graficos}
\usage{
generate_all_plots(x, verbose = TRUE)
}
\arguments{
\item{x}{Objeto easyml}

\item{verbose}{Mostrar progreso}
}
\value{
Lista con graficos y catalogo de figuras
}
\description{
Genera todos los graficos relevantes con metadatos para reportes cientificos.
Los graficos se almacenan con numeracion (Figura 1, Figura 2, etc.) y
nombres descriptivos para facilitar su referencia en reportes.
}
