% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_metrics}
\alias{generate_metrics}
\title{Generate model performance metrics.}
\usage{
generate_metrics(object)
}
\arguments{
\item{object}{TO BE EDITED.}
}
\value{
A list of matrixes, the generated model performance metrics.
}
\description{
Generate model performance metrics.
}
\seealso{
Other generate: \code{\link{generate_coefficients}},
  \code{\link{generate_predictions}},
  \code{\link{generate_variable_importances}}
}

