% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_model_performance}
\alias{generate_model_performance}
\title{Generate measures of model performance for a model.}
\usage{
generate_model_performance(object)
}
\arguments{
\item{object}{A list of class \code{easy_*}, where * is the name of the algorithm.}
}
\value{
A list of matrixes, the generated measures of model performance.
}
\description{
Generate measures of model performance for a model.
}
\seealso{
Other generate: \code{\link{generate_coefficients}},
  \code{\link{generate_predictions}},
  \code{\link{generate_variable_importances}}
}
