% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_predictions}
\alias{generate_predictions}
\title{Generate predictions.}
\usage{
generate_predictions(object)
}
\arguments{
\item{object}{TO BE EDITED.}
}
\value{
A list of matrixes, the generated predictions.
}
\description{
Generate predictions.
}
\seealso{
Other generate: \code{\link{generate_coefficients}},
  \code{\link{generate_model_performance}},
  \code{\link{generate_variable_importances}}
}

