% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_variable_importances}
\alias{generate_variable_importances}
\title{Generate variable importances.}
\usage{
generate_variable_importances(object)
}
\arguments{
\item{object}{TO BE EDITED.}
}
\value{
A data.frame, the generated variable importance scores.
}
\description{
Generate variable importances.
}
\seealso{
Other generate: \code{\link{generate_coefficients}},
  \code{\link{generate_model_performance}},
  \code{\link{generate_predictions}}
}

