% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_easyml.R
\name{get_figures_catalog}
\alias{get_figures_catalog}
\title{Obtener Catalogo de Figuras}
\usage{
get_figures_catalog(x)
}
\arguments{
\item{x}{Objeto easyml}
}
\value{
Data frame con el catalogo de figuras
}
\description{
Devuelve el catalogo de figuras generadas por easy_ml().
}
