% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet_extract_coefficients}
\alias{glmnet_extract_coefficients}
\title{Extract coefficients from a penalized regression model.}
\usage{
glmnet_extract_coefficients(results)
}
\arguments{
\item{results}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
Extract coefficients from a penalized regression model.
}

