% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet_predict_model}
\alias{glmnet_predict_model}
\title{Predict values for a penalized regression model.}
\usage{
glmnet_predict_model(results, newx = NULL)
}
\arguments{
\item{results}{TO BE EDITED.}

\item{newx}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
Predict values for a penalized regression model.
}

