% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isolate_dependent_variable}
\alias{isolate_dependent_variable}
\title{TO BE EDITED.}
\usage{
isolate_dependent_variable(.data = NULL, dependent_variable = NULL)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

