% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isolate_independent_variables}
\alias{isolate_independent_variables}
\title{TO BE EDITED.}
\usage{
isolate_independent_variables(.data, dependent_variable)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

