% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_cor_score}
\alias{measure_cor_score}
\title{Measure Correlation score.}
\usage{
measure_cor_score(x, y = NULL, use = "everything", method = c("pearson",
  "kendall", "spearman"))
}
\arguments{
\item{y_true}{A numeric vector; the ground truth (correct) target values.}

\item{y_pred}{A numeric vector; the estimated target values.}
}
\value{
A numeric vector of length one; the correlation metric.
}
\description{
Given the ground truth (correct) target values and the estimated target 
values, calculates the correlation metric.
}
\seealso{
Other measure: \code{\link{measure_area_under_curve}},
  \code{\link{measure_mean_squared_error}},
  \code{\link{measure_r2_score}}
}

