% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_correlation_score}
\alias{measure_correlation_score}
\title{Measure Pearsons Correlation Coefficient.}
\usage{
measure_correlation_score(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector; the ground truth (correct) target values.}

\item{y_pred}{A numeric vector; the estimated target values.}
}
\value{
A numeric vector of length one; the correlation metric.
}
\description{
Given the ground truth (correct) target values and the estimated target 
values, calculates the correlation metric.
}
\details{
See here for more information: \url{https://en.wikipedia.org/wiki/Pearson_correlation_coefficient}
}
\seealso{
Other measure: \code{\link{measure_auc_score}},
  \code{\link{measure_mse_score}},
  \code{\link{measure_r2_score}}
}
