% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_icc_score}
\alias{measure_icc_score}
\title{Measure Intraclass Correlation Coefficient (ICC(1,1)).}
\usage{
measure_icc_score(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector; the ground truth (correct) target values.}

\item{y_pred}{A numeric vector; the estimated target values.}
}
\value{
A numeric vector of length one; the ICC metric.
}
\description{
Given the ground truth (correct) target values and the estimated target 
values, calculates the ICC metric.
}
\details{
See here for more information: \url{https://en.wikipedia.org/wiki/Pearson_correlation_coefficient}
}
\seealso{
Other measure: \code{\link{measure_area_under_curve}},
  \code{\link{measure_cor_score}},
  \code{\link{measure_mean_squared_error}},
  \code{\link{measure_r2_score}}
}

