% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_r2_score}
\alias{measure_r2_score}
\title{Measure Coefficient of Determination (R^2 Score).}
\usage{
measure_r2_score(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector; the ground truth (correct) target values.}

\item{y_pred}{A numeric vector; the estimated target values.}
}
\value{
A numeric vector of length one; the R^2 metric.
}
\description{
Given the ground truth (correct) target values and the estimated target 
values, calculates the the R^2 metric.
}
\details{
See here for more information: \url{https://en.wikipedia.org/wiki/Coefficient_of_determination}
}
\seealso{
Other measure: \code{\link{measure_auc_score}},
  \code{\link{measure_correlation_score}},
  \code{\link{measure_mse_score}}
}
