% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_ml.R
\name{plot.causalml}
\alias{plot.causalml}
\title{Plot method for causalml}
\usage{
\method{plot}{causalml}(x, type = "all", ...)
}
\arguments{
\item{x}{Objeto de clase causalml.}

\item{type}{Tipo de grafico: "all", "treatment", "outcome", "ps", "overlap",
"love", "cate", "heterogeneity", "importance", "ate_forest", "meta",
"evalue", "balance".}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Plot method for causalml
}
