% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_ml.R
\name{plot.reinforcementml}
\alias{plot.reinforcementml}
\title{Plot method for reinforcementml}
\usage{
\method{plot}{reinforcementml}(x, type = "all", ...)
}
\arguments{
\item{x}{Objeto de clase reinforcementml.}

\item{type}{Tipo de grafico: "all", "reward", "regret", "actions",
"qvalues", "returns", "comparison", "posteriors".}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Plot method for reinforcementml
}
