% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_learning_curve.R
\name{plot_all.ml_sample_size}
\alias{plot_all.ml_sample_size}
\title{Plot All Graphs for Sample Size Results}
\usage{
\method{plot_all}{ml_sample_size}(x, ...)
}
\arguments{
\item{x}{Object of class ml_sample_size}

\item{...}{Additional arguments}
}
\description{
Displays learning curve and distribution plots side by side.
}
