% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_easyml.R
\name{plot_all}
\alias{plot_all}
\title{Graficar todos los resultados}
\usage{
plot_all(x, ...)
}
\arguments{
\item{x}{Objeto easyml}

\item{...}{Argumentos adicionales}
}
\value{
Panel de graficos
}
\description{
Genera un panel con todos los graficos relevantes.
}
