% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coefficients_processed}
\alias{plot_coefficients_processed}
\title{TO BE EDITED.}
\usage{
plot_coefficients_processed(coefficients_processed)
}
\arguments{
\item{coefficients_processed}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

