% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_learning_curve.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Graficar distribucion de metricas por tamano de muestra}
\usage{
plot_distribution(x, ...)
}
\arguments{
\item{x}{Objeto de clase ml_sample_size}

\item{...}{Argumentos adicionales}
}
\description{
Muestra boxplots de la distribucion de la metrica para cada tamano de muestra
}
