% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metrics_gaussian_mean_squared_error}
\alias{plot_metrics_gaussian_mean_squared_error}
\title{Plot mean squared error metrics.}
\usage{
plot_metrics_gaussian_mean_squared_error(mses)
}
\arguments{
\item{mses}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
Plot mean squared error metrics.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_metrics_binomial_area_under_curve}},
  \code{\link{plot_metrics_gaussian_cor_score}},
  \code{\link{plot_metrics_gaussian_r2_score}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_variable_importances}}
}

