% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metrics_gaussian_r2_score}
\alias{plot_metrics_gaussian_r2_score}
\title{Plot coefficient of determination (R^2) metrics.}
\usage{
plot_metrics_gaussian_r2_score(r2_scores)
}
\arguments{
\item{r2_scores}{A vector, the coefficient of determination (R^2) metrics to be plotted as a histogram.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
Plot coefficient of determination (R^2) metrics.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_metrics_binomial_area_under_curve}},
  \code{\link{plot_metrics_gaussian_cor_score}},
  \code{\link{plot_metrics_gaussian_mean_squared_error}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_variable_importances}}
}

