% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metrics_histogram}
\alias{plot_metrics_histogram}
\title{Plot histogram of metrics.}
\usage{
plot_metrics_histogram(x, name)
}
\arguments{
\item{mses}{A vector, the mean squared error metrics to be plotted as a histogram.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
Plot histogram of metrics.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_metrics_binomial_auc_score}},
  \code{\link{plot_metrics_gaussian_correlation_score}},
  \code{\link{plot_metrics_gaussian_mse_score}},
  \code{\link{plot_metrics_gaussian_r2_score}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_roc_curve}},
  \code{\link{plot_variable_importances_processed}}
}

