% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_model_performance_gaussian_mse_score}
\alias{plot_model_performance_gaussian_mse_score}
\title{Plot histogram of the mean squared error metrics.}
\usage{
plot_model_performance_gaussian_mse_score(x)
}
\arguments{
\item{x}{A vector, the mean squared error metrics to be plotted as a histogram.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
This function plots a histogram of the mean squared error metrics.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_model_performance_binomial_auc_score}},
  \code{\link{plot_model_performance_gaussian_correlation_score}},
  \code{\link{plot_model_performance_gaussian_r2_score}},
  \code{\link{plot_model_performance_histogram}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_roc_curve}},
  \code{\link{plot_variable_importances_processed}}
}
