% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_model_performance_histogram}
\alias{plot_model_performance_histogram}
\title{Plot histogram of measures of model performance.}
\usage{
plot_model_performance_histogram(x, name)
}
\arguments{
\item{x}{A vector, the mean squared error metrics to be plotted as a histogram.}

\item{name}{A character vector of length one, the name of the metric.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
Given a numeric vector will plot a histogram 
for any number of measures of model performance.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_model_performance_binomial_auc_score}},
  \code{\link{plot_model_performance_gaussian_correlation_score}},
  \code{\link{plot_model_performance_gaussian_mse_score}},
  \code{\link{plot_model_performance_gaussian_r2_score}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_roc_curve}},
  \code{\link{plot_variable_importances_processed}}
}
