% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_predictions_binomial}
\alias{plot_predictions_binomial}
\title{TO BE EDITED.}
\usage{
plot_predictions_binomial(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

