% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_predictions_gaussian}
\alias{plot_predictions_gaussian}
\title{Plot gaussian predictions.}
\usage{
plot_predictions_gaussian(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
Plot gaussian predictions.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_metrics_binomial_area_under_curve}},
  \code{\link{plot_metrics_gaussian_cor_score}},
  \code{\link{plot_metrics_gaussian_mean_squared_error}},
  \code{\link{plot_metrics_gaussian_r2_score}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_variable_importances_processed}}
}

