% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{plot_residual_diagnostics}
\alias{plot_residual_diagnostics}
\title{Gráficos de Diagnóstico de Residuos}
\usage{
plot_residual_diagnostics(residual_analysis)
}
\description{
Gráficos de Diagnóstico de Residuos
}
