% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc_auc.R
\name{plot_roc_auc}
\alias{plot_roc_auc}
\title{TO BE EDITED.}
\usage{
plot_roc_auc(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

