% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{TO BE EDITED.}
\usage{
plot_roc_curve(y_true, y_pred)

plot_roc_curve(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}

\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
TO BE EDITED.

TO BE EDITED.
}
\description{
TO BE EDITED.

TO BE EDITED.
}

