% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretability.R
\name{plot_shap_dependence}
\alias{plot_shap_dependence}
\title{Grafico SHAP para una variable}
\usage{
plot_shap_dependence(interpret_result, variable, test_data)
}
\description{
Grafico SHAP para una variable
}
