% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_variable_importances}
\alias{plot_variable_importances}
\title{Plot variable importances.}
\usage{
plot_variable_importances(variable_importances)
}
\arguments{
\item{variable_importances}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
Plot variable importances.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_metrics_binomial_area_under_curve}},
  \code{\link{plot_metrics_gaussian_cor_score}},
  \code{\link{plot_metrics_gaussian_mean_squared_error}},
  \code{\link{plot_metrics_gaussian_r2_score}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}}
}

