% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_ml.R
\name{predict.reinforcementml}
\alias{predict.reinforcementml}
\title{Predict method for reinforcementml}
\usage{
\method{predict}{reinforcementml}(object, state = NULL, algorithm = NULL, ...)
}
\arguments{
\item{object}{Objeto de clase reinforcementml.}

\item{state}{Estado actual (numerico para MDP, vector de contexto para bandits contextuales).}

\item{algorithm}{Algoritmo a usar. Default: el mejor.}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Predict method for reinforcementml
}
