% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyml.R
\name{predict.supervisedml}
\alias{predict.supervisedml}
\title{Predecir con modelo supervised_ml}
\usage{
\method{predict}{supervisedml}(object, new_data, type = "class", ...)
}
\arguments{
\item{object}{Objeto de clase supervisedml.}

\item{new_data}{Data frame con nuevos datos para predecir.}

\item{type}{Tipo de prediccion: "class" (default) o "prob" (probabilidades, solo clasificacion).}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Predecir con modelo supervised_ml
}
