% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_helpers.R
\name{predict.unsupervisedml}
\alias{predict.unsupervisedml}
\title{Asignar clusters a nuevos datos}
\usage{
\method{predict}{unsupervisedml}(result, newdata, algorithm = NULL, ...)
}
\arguments{
\item{result}{Objeto de clase \code{unsupervisedml}.}

\item{newdata}{Data frame con nuevos datos.}

\item{algorithm}{Algoritmo a usar (default: mejor).}
}
\value{
Vector de labels de cluster.
}
\description{
Asignar clusters a nuevos datos
}
