% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avNNet.R
\name{predict_model.easy_avNNet}
\alias{predict_model.easy_avNNet}
\title{Predict values for a neural network model.}
\usage{
\method{predict_model}{easy_avNNet}(object, newx = NULL)
}
\arguments{
\item{object}{A list of class \code{easy_avNNet}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values using the new data.
}
\description{
Predict values for a neural network model.
}
