% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinternet.R
\name{predict_model.easy_glinternet}
\alias{predict_model.easy_glinternet}
\title{Predict values for a penalized regression model.}
\usage{
\method{predict_model}{easy_glinternet}(object, newx = NULL)
}
\arguments{
\item{newx}{A data.frame, the new data to use for predictions.}

\item{results}{The results of \code{\link{fit_model}}.}
}
\value{
A vector, the predicted values for a penalized regression model using the new data.
}
\description{
Predict values for a penalized regression model.
}

