% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{predict_model.easy_glmnet}
\alias{predict_model.easy_glmnet}
\title{Predict values for a penalized regression model.}
\usage{
\method{predict_model}{easy_glmnet}(results, newx = NULL)
}
\arguments{
\item{results}{The results of \code{\link{fit_model}}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values for a penalized regression model using the new data.
}
\description{
Predict values for a penalized regression model.
}

