% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.R
\name{predict_model.easy_neural_network}
\alias{predict_model.easy_neural_network}
\title{Predict values for a neural network model.}
\usage{
\method{predict_model}{easy_neural_network}(object, newx = NULL)
}
\arguments{
\item{object}{A list of class \code{easy_neural_network}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values using the new data.
}
\description{
This function wraps the procedure for predicting values from 
a neural network model and makes it accessible 
to the easyml core framework.
}
