% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{predict_model.easy_random_forest}
\alias{predict_model.easy_random_forest}
\title{Predict values for a random forest regression model.}
\usage{
\method{predict_model}{easy_random_forest}(object, newx = NULL)
}
\arguments{
\item{object}{A list of class \code{easy_random_forest}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values using the new data.
}
\description{
This function wraps the procedure for predicting values from 
a random forest model and makes it accessible 
to the easyml core framework.
}
