% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{predict_model.easy_support_vector_machine}
\alias{predict_model.easy_support_vector_machine}
\title{Predict values for a support vector machine regression model.}
\usage{
\method{predict_model}{easy_support_vector_machine}(object, newx = NULL)
}
\arguments{
\item{object}{A list of class \code{easy_support_vector_machine}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values using the new data.
}
\description{
This function wraps the procedure for predicting values from 
a support vector machine model and makes it accessible 
to the easyml core framework.
}
