% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{predict_model}
\alias{predict_model}
\title{Predict model.}
\usage{
predict_model(object, newx = NULL)
}
\arguments{
\item{object}{A list of class \code{easy_*}, where * is the name of the algorithm.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values using the new data.
}
\description{
The generic function  for generating predictions from a model within 
the easyml core framework. Users can create their own 
predict_model function for their own class of model.
}
