% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prep_feature_selection}
\alias{prep_feature_selection}
\title{Feature Selection con Boruta}
\usage{
prep_feature_selection(data, target, seed = 2024, verbose = TRUE)
}
\description{
Feature Selection con Boruta
}
