% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_split.R
\name{print.block_split}
\alias{print.block_split}
\title{Print method for block_split objects}
\usage{
\method{print}{block_split}(x, ...)
}
\arguments{
\item{x}{A block_split object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for block_split objects
}
