% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_ml.R
\name{print.causalml}
\alias{print.causalml}
\title{Print method for causalml}
\usage{
\method{print}{causalml}(x, ...)
}
\arguments{
\item{x}{Objeto de clase causalml.}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Print method for causalml
}
