% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_split.R
\name{print.diversity_split}
\alias{print.diversity_split}
\title{Print method for diversity_split objects}
\usage{
\method{print}{diversity_split}(x, ...)
}
\description{
Print method for diversity_split objects
}
