% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyml.R
\name{print.supervisedml}
\alias{print.supervisedml}
\title{Print method for supervisedml}
\usage{
\method{print}{supervisedml}(x, ...)
}
\arguments{
\item{x}{Objeto de clase supervisedml.}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Print method for supervisedml
}
