% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_easyml.R
\name{print_figures_catalog}
\alias{print_figures_catalog}
\title{Imprimir Catalogo de Figuras}
\usage{
print_figures_catalog(x)
}
\arguments{
\item{x}{Objeto easyml}
}
\value{
NULL (invisible)
}
\description{
Imprime el catalogo de figuras para el usuario.
}
