% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_coefficients}
\alias{process_coefficients}
\title{Process coefficients.}
\usage{
process_coefficients(coefs, survival_rate_cutoff = 0.05)
}
\arguments{
\item{coefs}{The replicated coefficients.}

\item{survival_rate_cutoff}{A numeric vector of length one; for \code{\link{easy_glmnet}}, specifies the minimal threshold (as a percentage) a coefficient must appear out of n_samples. Defaults to 0.05.}
}
\value{
TO BE EDITED.
}
\description{
Process coefficients.
}
\seealso{
Other utils: \code{\link{correlation_test}},
  \code{\link{reduce_cores}},
  \code{\link{remove_variables}}
}

