% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_data}
\alias{process_data}
\title{TO BE EDITED.}
\usage{
process_data(.df, dependent_variable = NULL, exclude_variables = NULL)
}
\arguments{
\item{dependent_variable}{TO BE EDITED.}

\item{exclude_variables}{TO BE EDITED.}

\item{data}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

