% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_variable_importances}
\alias{process_variable_importances}
\title{Process variable importances.}
\usage{
process_variable_importances(variable_importances)
}
\arguments{
\item{variable_importances}{A data.frame, the replicated coefficients.}
}
\value{
A data.frame; the replicated variable importances processed for easy plotting.
}
\description{
Takes the variable importances returned by the 
generate_variable_importances function and prepares 
the variable importances for plotting.
}
\seealso{
Other utils: \code{\link{correlation_test}},
  \code{\link{process_coefficients}},
  \code{\link{reduce_cores}},
  \code{\link{remove_variables}}
}
