% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_glmnet.R
\name{quick_glmnet}
\alias{quick_glmnet}
\title{Quickly and easily run penalized linear or logistic regression on a dataset.}
\usage{
quick_glmnet(dat_path = NULL, dependentVar = NULL, depCate = NULL,
  excludeVar = NULL, categoricalVar = NULL, numIterations = 1000,
  header = T, outOfSample = T, splitBy = 3, whichSeq = 1, myAlpha = 1,
  nFolds = 5, survivalRate_cutoff = 0.05)
}
\arguments{
\item{dat_path}{path to a data file}

\item{dependentVar}{dependent variable}

\item{depCate}{category of DV: "binary" or "continuous"}

\item{excludeVar}{any variables to exlcude?}

\item{categoricalVar}{any categorical variables?}

\item{numIterations}{Default=1000. Number of iterations}

\item{header}{Does the data file have a header? T or F}

\item{outOfSample}{out-of-sample predictions? T or F}

\item{splitBy}{Default=3. Use 1/splitBy e.g. 33 percent of data as a test set}

\item{whichSeq}{Default=1. Which sequence to use for a test set? 1, 2, or 3}

\item{myAlpha}{Default=1. myAlpha=1 for LASSO}

\item{nFolds}{Default=5. k-fold}

\item{survivalRate_cutoff}{cutoff to be included in the beta matrix}
}
\value{
TO BE EDITED.
}
\description{
Quickly and easily run penalized linear or logistic regression on a dataset.
}

