% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest_extract_variable_importances}
\alias{random_forest_extract_variable_importances}
\title{Extract variable importance scores from a random forest model.}
\usage{
random_forest_extract_variable_importances(results)
}
\arguments{
\item{results}{The results of \code{\link{random_forest_fit_model_gaussian}} or \code{\link{random_forest_fit_model_gaussian}}.}
}
\value{
A data.frame of replicated random forest variable importance scores.
}
\description{
Extract variable importance scores from a random forest model.
}

