% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce_cores}
\alias{reduce_cores}
\title{Reduce number of cores.}
\usage{
reduce_cores(n_core, cpu_count = NULL)
}
\arguments{
\item{n_core}{An integer vector of length one; specifies the number of cores to use for this analysis. Currently only works on Mac OSx and Unix/Linux systems. Defaults to 1L.}

\item{cpu_count}{An integer vector of length one; specifies the number of cores potentially available to use for this analysis. Currently only works on Mac OSx and Unix/Linux systems. Defaults to 1L.}
}
\value{
An integer vector of length one; specifies the number of cores to use for this analysis.
}
\description{
This function takes the number of cores specified by the user and
reduces it to the maximum number of cores supported by the computer.
}
\seealso{
Other utils: \code{\link{correlation_test}},
  \code{\link{process_coefficients}},
  \code{\link{process_variable_importances}},
  \code{\link{remove_variables}}
}
