% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_variables}
\alias{remove_variables}
\title{Remove variables from a dataset.}
\usage{
remove_variables(.data = NULL, exclude_variables = NULL)
}
\arguments{
\item{.data}{A data.frame; the data to be analyzed.}

\item{exclude_variables}{A character vector; the variables from the data set to exclude. Defaults to NULL.}
}
\value{
A data.frame; the data to be analyzed.
}
\description{
This utility function removes variables from a data.frame.
}
\seealso{
Other utils: \code{\link{correlation_test}},
  \code{\link{process_coefficients}},
  \code{\link{process_variable_importances}},
  \code{\link{reduce_cores}}
}
