% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_aucs}
\alias{replicate_aucs}
\title{TO BE EDITED.}
\usage{
replicate_aucs(fit_model, predict_model, sampler, preprocessor, X, y,
  categorical_variables = NULL, n_divisions = 1000, n_iterations = 100,
  progress_bar = TRUE, n_core = 1, ...)
}
\arguments{
\item{fit_model}{TO BE EDITED.}

\item{predict_model}{TO BE EDITED.}

\item{X}{TO BE EDITED.}

\item{y}{TO BE EDITED.}

\item{n_divisions}{TO BE EDITED.}

\item{n_iterations}{TO BE EDITED.}

\item{progress_bar}{TO BE EDITED.}

\item{predict_model}{TO BE EDITED.}

\item{parallel}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

