% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_coefficients}
\alias{replicate_coefficients}
\title{Replicate coefficients.}
\usage{
replicate_coefficients(object)
}
\arguments{
\item{object}{TO BE EDITED.}
}
\value{
A data.frame, the replicated penalized regression model coefficients.
}
\description{
Replicate coefficients.
}
\seealso{
Other replicate: \code{\link{replicate_metrics}},
  \code{\link{replicate_predictions}},
  \code{\link{replicate_variable_importances}}
}

