% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_metrics}
\alias{replicate_metrics}
\title{Replicate metrics.}
\usage{
replicate_metrics(object)
}
\arguments{
\item{object}{TO BE EDITED.}
}
\value{
A list of matrixes, the replicated metrics.
}
\description{
Replicate metrics.
}
\seealso{
Other replicate: \code{\link{replicate_coefficients}},
  \code{\link{replicate_predictions}},
  \code{\link{replicate_variable_importances}}
}

