% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_predictions}
\alias{replicate_predictions}
\title{TO BE EDITED.}
\usage{
replicate_predictions(fit_model, predict_model, preprocessor, X_train, y_train,
  X_test, categorical_variables = NULL, n_samples = 1000,
  progress_bar = TRUE, n_core = 1, ...)
}
\arguments{
\item{fit_model}{TO BE EDITED.}

\item{predict_model}{TO BE EDITED.}

\item{X_train}{TO BE EDITED.}

\item{y_train}{TO BE EDITED.}

\item{X_test}{TO BE EDITED.}

\item{n_samples}{TO BE EDITED.}

\item{progress_bar}{TO BE EDITED.}

\item{parallel}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

