% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_variable_importances}
\alias{replicate_variable_importances}
\title{Replicate variable importances.}
\usage{
replicate_variable_importances(object)
}
\arguments{
\item{object}{TO BE EDITED.}
}
\value{
A data.frame, the replicated variable importance scores.
}
\description{
Replicate variable importances.
}
\seealso{
Other replicate: \code{\link{replicate_coefficients}},
  \code{\link{replicate_metrics}},
  \code{\link{replicate_predictions}}
}

