% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_simple_train_test_split}
\alias{resample_simple_train_test_split}
\title{Train test split.}
\usage{
resample_simple_train_test_split(X, y, train_size = 0.667,
  random_state = NULL)
}
\arguments{
\item{X}{TO BE EDITED.}

\item{y}{A numeric vector with two classes, 0 and 1.}

\item{train_size}{The proportoin of data in training set.}

\item{random_state}{An integer to seed the random number generator.}
}
\value{
A boolean vector of length n_obs where TRUE represents that observation should be in the train set.
}
\description{
This will split the data into train and test.
}

