% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_equal_proportion}
\alias{sample_equal_proportion}
\title{Sample in equal proportion.}
\usage{
sample_equal_proportion(y, proportion = 0.667, seed = 1)
}
\arguments{
\item{y}{A numeric vector with two classes, 0 and 1.}

\item{proportion}{The proportoin of data in training set.}

\item{seed}{An integer to seed the random number generator.}
}
\value{
A boolean vector of length n_obs where TRUE represents that observation should be in the train set.
}
\description{
This will sample in equal proportion.
}

