% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_categorical_variables}
\alias{set_categorical_variables}
\title{TO BE EDITED.}
\usage{
set_categorical_variables(column_names, categorical_variables = NULL)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

