% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_coefficients_boolean}
\alias{set_coefficients_boolean}
\title{Set coefficients boolean.}
\usage{
set_coefficients_boolean(algorithm)
}
\arguments{
\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}
}
\value{
A logical vector of length one; whether coefficients should be replicated for this analyis or not.
}
\description{
Set coefficients boolean.
}

