% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_cores}
\alias{set_cores}
\title{Set cores.}
\usage{
set_cores(n_core)
}
\arguments{
\item{n_core}{An integer vector of length one; specifies the number of cores to use for this analysis. Currenly only works on Mac OSx and Unix/Linux systems. Defaults to 1L.}
}
\value{
NULL.
}
\description{
Please note this affects global state and sets the number of cores by running \code{options(mc.cores = n_core)}.
}

