% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_dependent_variable}
\alias{set_dependent_variable}
\title{Set dependent variable.}
\usage{
set_dependent_variable(.data, dependent_variable)
}
\arguments{
\item{.data}{A data.frame; the data to be analyzed.}

\item{dependent_variable}{A character vector of length one; the dependent variable for this analysis.}
}
\value{
TO BE EDITED.
}
\description{
Set dependent variable.
}

