% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_fit_model}
\alias{set_fit_model}
\title{Set fit model function.}
\usage{
set_fit_model(algorithm, family)
}
\arguments{
\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}

\item{family}{A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".}
}
\value{
TO BE EDITED.
}
\description{
Sets the function responsible for fitting a model to the data.
}

