% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_independent_variables}
\alias{set_independent_variables}
\title{Set independent variablex.}
\usage{
set_independent_variables(.data, dependent_variable)
}
\arguments{
\item{.data}{A data.frame; the data to be analyzed.}

\item{dependent_variable}{A character vector of length one; the dependent variable for this analysis.}
}
\value{
TO BE EDITED.
}
\description{
Set independent variablex.
}

