% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_looper}
\alias{set_looper}
\title{Set looper.}
\usage{
set_looper(progress_bar = FALSE, n_core = 1)
}
\arguments{
\item{progress_bar}{A logical vector of length one; specifies whether to display a progress bar during calculations. Defaults to TRUE.}

\item{n_core}{An integer vector of length one; specifies the number of cores to use for this analysis. Currenly only works on Mac OSx and Unix/Linux systems. Defaults to 1L.}
}
\value{
TO BE EDITED.
}
\description{
This function decides which looper (a functional like lapply) to run. Please note this affects global state and sets the number of cores by ultimately running \code{options(mc.cores = n_core)}.
}

