% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_looper_}
\alias{set_looper_}
\title{Set looper.}
\usage{
set_looper_(progress_bar = FALSE, parallel = FALSE)
}
\arguments{
\item{progress_bar}{A logical vector of length one; specifies whether to display a progress bar during calculations. Defaults to FALSE.}

\item{parallel}{A logical vector of length one; specifies whether to run calculations in parallel. Defaults to FALSE.}
}
\value{
TO BE EDITED.
}
\description{
This function decides which looper (a functional like lapply) to run. This function does not affect global state.
}

