% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_measure}
\alias{set_measure}
\title{Set measure function.}
\usage{
set_measure(measure = NULL, algorithm, family)
}
\arguments{
\item{measure}{A function; the function for measuring the results. Defaults to NULL.}

\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}

\item{family}{A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".}
}
\value{
TO BE EDITED.
}
\description{
Sets the function responsible for measuring the results.
}

