% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_metrics_boolean}
\alias{set_metrics_boolean}
\title{Set metrics boolean.}
\usage{
set_metrics_boolean(algorithm)
}
\arguments{
\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}
}
\value{
A logical vector of length one; whether metrics should be replicated for this analyis or not.
}
\description{
Set metrics boolean.
}

