% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_parallel}
\alias{set_parallel}
\title{Set parallel.}
\usage{
set_parallel(n_core)
}
\arguments{
\item{n_core}{An integer vector of length one; specifies the number of cores to use for this analysis. Currenly only works on Mac OSx and Unix/Linux systems. Defaults to 1L.}
}
\value{
A logical vector of length one; whether analyis should be run in parallel or not.
}
\description{
Set parallel.
}

