% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_plot_metrics}
\alias{set_plot_metrics}
\title{Set plot metrics function.}
\usage{
set_plot_metrics(measure)
}
\arguments{
\item{measure}{A function; the function for measuring the results. Defaults to NULL.}
}
\value{
TO BE EDITED.
}
\description{
Sets the function responsible for plotting the metrics generated from the predictions generated from a fitted model.
}

