% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_plot_model_performance}
\alias{set_plot_model_performance}
\title{Set plot model performance function.}
\usage{
set_plot_model_performance(measure)
}
\arguments{
\item{measure}{A function; the function for measuring the results. Defaults to NULL.}
}
\value{
TA function; the function for plotting the measures of model performance generated from the predictions generated from a fitted model.
}
\description{
Sets the function responsible for plotting the measures of model performance generated from the predictions generated from a fitted model.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}}, \code{\link{set_parallel}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}}
}
