% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_preprocess}
\alias{set_preprocess}
\title{Set preprocess function.}
\usage{
set_preprocess(preprocess = NULL)
}
\arguments{
\item{preprocess}{A function; the function for preprocessing the data. Defaults to NULL.}
}
\value{
TO BE EDITED.
}
\description{
Sets the function responsible for preprocessing the data.
}

