% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_random_state}
\alias{set_random_state}
\title{Set random state.}
\usage{
set_random_state(random_state = NULL)
}
\arguments{
\item{random_state}{An integer vector of length one; specifies the seed to be used for the analysis. Defaults to NULL.}
}
\value{
TO BE EDITED.
}
\description{
Sets the random state to a specific seed. Please note this function affects global state.
}

