% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_resample}
\alias{set_resample}
\title{Set resample function.}
\usage{
set_resample(resample = NULL, family = NULL)
}
\arguments{
\item{resample}{A function; the function for resampling the data. Defaults to NULL.}

\item{family}{A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".}
}
\value{
TO BE EDITED.
}
\description{
Sets the function responsible for resampling the data.
}

