% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_ml.R
\name{summary.causalml}
\alias{summary.causalml}
\title{Summary method for causalml}
\usage{
\method{summary}{causalml}(object, ...)
}
\arguments{
\item{object}{Objeto de clase causalml.}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Summary method for causalml
}
