% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyml.R
\name{summary.supervisedml}
\alias{summary.supervisedml}
\title{Resumen del modelo supervised_ml}
\usage{
\method{summary}{supervisedml}(object, ...)
}
\arguments{
\item{object}{Objeto de clase supervisedml.}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Resumen del modelo supervised_ml
}
