% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_ml.R
\name{summary.unsupervisedml}
\alias{summary.unsupervisedml}
\title{Summary method for unsupervisedml}
\usage{
\method{summary}{unsupervisedml}(object, ...)
}
\arguments{
\item{object}{Objeto de clase unsupervisedml.}

\item{...}{Argumentos adicionales (no usados).}
}
\description{
Summary method for unsupervisedml
}
