% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_verbose.R
\name{supervised_ml_capture}
\alias{supervised_ml_capture}
\title{Capturar Verbose Output}
\usage{
supervised_ml_capture(...)
}
\arguments{
\item{...}{Argumentos para supervised_ml()}
}
\value{
Lista con resultados y verbose capturado
}
\description{
Ejecuta supervised_ml capturando todo el output verbose para exportacion.
}
