% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{support_vector_machine_fit_model_gaussian}
\alias{support_vector_machine_fit_model_gaussian}
\title{TO BE EDITED.}
\usage{
support_vector_machine_fit_model_gaussian(X, y, ...)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

