% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{support_vector_machine_predict_model}
\alias{support_vector_machine_predict_model}
\title{TO BE EDITED.}
\usage{
support_vector_machine_predict_model(results, newx)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

