% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_coefficients}
\alias{bootstrap_coefficients}
\title{TO BE EDITED.}
\usage{
bootstrap_coefficients(fit_model, extract_coefficients, X, y,
  n_samples = 1000)
}
\arguments{
\item{fit_model}{TO BE EDITED.}

\item{extract_coefficients}{TO BE EDITED.}

\item{X}{TO BE EDITED.}

\item{y}{TO BE EDITED.}

\item{n_samples}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

