% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_predictions}
\alias{bootstrap_predictions}
\title{TO BE EDITED.}
\usage{
bootstrap_predictions(estimator, X_train, y_train, X_test, n_samples = 1000)
}
\arguments{
\item{estimator}{TO BE EDITED.}

\item{X_train}{TO BE EDITED.}

\item{y_train}{TO BE EDITED.}

\item{X_test}{TO BE EDITED.}

\item{n_samples}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

