% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cocaine_dependence}
\alias{cocaine_dependence}
\title{Cocaine data.}
\format{Data frame with columns
\describe{
\item{subject}{Subject ID}
\item{diagnosis}{Diagnosis status}
\item{age}{Age of Subject}
\item{male}{Gender of subject}
\item{edu_yrs}{Education of subject in years}
\item{imt_comm_errors}{Immediate Memory Task commission errors}
\item{imt_omis_errors}{Immediate Memory Task omission errors}
\item{a_imt}{Immediate Memory Task non-parametric discriminability}
\item{b_d_imt}{Immediate Memory Task response bias }
\item{stop_ssrt}{Stop-Signal Task stop-signal reaction time}
\item{lnk_adjdd}{log(Adjusting Delay-Discounting Task discounting rate)}
\item{lkhat_kirby}{log(Kirby Monetary-Choice Delay-Discounting Questionnaire)}
\item{revlr_per_errors}{Probabilistic Reversal-Learning Task errors}
\item{bis_attention}{Barratt Impulsiveness Scale attentional impulsivity}
\item{bis_motor}{Barratt Impulsiveness Scale motor impulsivity}
\item{bis_nonpL}{Barratt Impulsiveness Scale nonplanning impulsivity}
\item{igt_total}{Iowa Gambling Task total score}
}}
\source{
https://github.com/CCS-Lab/easyml/blob/master/Python/datasets/cocaine_dependence.csv
}
\usage{
cocaine_dependence
}
\description{
Cocaine data.
}
\examples{
data("cocaine_dependence", package = "easyml")
}
\seealso{
Other data: \code{\link{prostate}}
}
\keyword{datasets}
