% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{easy_random_forest}
\alias{easy_random_forest}
\title{Easily build and evaluate a random forest regression model.}
\usage{
easy_random_forest(.data, dependent_variable, family = "gaussian",
  resample = NULL, preprocess = NULL, measure = NULL,
  exclude_variables = NULL, categorical_variables = NULL,
  train_size = 0.667, n_samples = 1000L, n_divisions = 1000L,
  n_iterations = 10L, random_state = NULL, progress_bar = TRUE,
  n_core = 1L, ...)
}
\arguments{
\item{.data}{A data.frame; the data to be analyzed.}

\item{dependent_variable}{A character vector of length one; the dependent variable for this analysis.}

\item{family}{A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".}

\item{resample}{A function; the function for resampling the data. Defaults to NULL.}

\item{preprocess}{A function; the function for preprocessing the data. Defaults to NULL.}

\item{measure}{A function; the function for measuring the results. Defaults to NULL.}

\item{exclude_variables}{A character vector; the variables from the data set to exclude. Defaults to NULL.}

\item{categorical_variables}{A character vector; the variables that are categorical. Defaults to NULL.}

\item{train_size}{A numeric vector of length one; specifies what proportion of the data should be used for the training data set. Defaults to 0.667.}

\item{n_samples}{An integer vector of length one; specifies the number of times the coefficients and predictions should be replicated. Defaults to 1000L.}

\item{n_divisions}{An integer vector of length one; specifies the number of times the data should be divided when replicating the error metrics. Defaults to 1000L.}

\item{n_iterations}{An integer vector of length one; during each division, specifies the number of times the predictions should be replicated. Defaults to 10L.}

\item{random_state}{An integer vector of length one; specifies the seed to be used for the analysis. Defaults to NULL.}

\item{progress_bar}{A logical vector of length one; specifies whether to display a progress bar during calculations. Defaults to TRUE.}

\item{n_core}{An integer vector of length one; specifies the number of cores to use for this analysis. Currenly only works on Mac OSx and Unix/Linux systems. Defaults to 1L.}

\item{...}{Arguments to be passed to \code{\link[randomForest]{randomForest}}. See that function's documentation for more details.}
}
\value{
TO BE EDITED.
}
\description{
Easily build and evaluate a random forest regression model.
}
\examples{
library(easyml) # https://github.com/CCS-Lab/easyml

# Gaussian
data("prostate", package = "easyml")
results <- easy_random_forest(prostate, "lpsa", 
                              n_samples = 10L, 
                              n_divisions = 10L, 
                              n_iterations = 2L, 
                              random_state = 12345L, n_core = 1L)

# Binomial
data("cocaine_dependence", package = "easyml")
results <- easy_random_forest(cocaine_dependence, "diagnosis", 
                              family = "binomial", 
                              exclude_variables = c("subject"),
                              categorical_variables = c("male"),
                              n_samples = 10L, 
                              n_divisions = 10L, 
                              n_iterations = 2L, 
                              random_state = 12345L, n_core = 1L)
}

