% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinternet.R
\name{extract_coefficients.easy_glinternet}
\alias{extract_coefficients.easy_glinternet}
\title{Extract coefficients from a penalized regression model with interactions.}
\usage{
\method{extract_coefficients}{easy_glinternet}(object)
}
\arguments{
\item{object}{A list of class \code{easy_glinternet}.}
}
\value{
A data.frame, the replicated penalized regression with interactions coefficients.
}
\description{
Extract coefficients from a penalized regression model with interactions.
}

