% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{extract_coefficients.easy_glmnet}
\alias{extract_coefficients.easy_glmnet}
\title{Extract coefficients from a penalized regression model.}
\usage{
\method{extract_coefficients}{easy_glmnet}(object)
}
\arguments{
\item{object}{A list of class \code{easy_glmnet}.}
}
\value{
A data.frame, the replicated penalized regression coefficients.
}
\description{
Extract coefficients from a penalized regression model.
}
