% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{extract_variable_importances}
\alias{extract_variable_importances}
\title{Extract variable importances.}
\usage{
extract_variable_importances(object)
}
\arguments{
\item{object}{A list of class \code{easy_*}, where * is the name of the algorithm.}
}
\value{
A data.frame, the generated random forest variable importance scores.
}
\description{
The generic function  for extracting variable importances from
a model within the easyml core framework, if such an
operation is applicable for that model. Users can create their own 
extract_variable_importances function for their own class of model.
}
