% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinternet.R
\name{fit_model.easy_glinternet}
\alias{fit_model.easy_glinternet}
\title{Fit a penalized regression model with interactions.}
\usage{
\method{fit_model}{easy_glinternet}(object)
}
\arguments{
\item{object}{A list of class \code{easy_glinternet}.}
}
\value{
A list of class \code{easy_glinternet}.
}
\description{
This function wraps the procedure for fitting a 
glinternet model and makes it accessible 
to the easyml core framework.
}
