% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{fit_model.easy_glmnet}
\alias{fit_model.easy_glmnet}
\title{Fit penalized regression model.}
\usage{
\method{fit_model}{easy_glmnet}(object)
}
\arguments{
\item{object}{TO BE EDITED.}
}
\value{
A list, the model and the cross validated model.
}
\description{
Fit penalized regression model.
}

