% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{fit_model.easy_glmnet}
\alias{fit_model.easy_glmnet}
\title{Fit a penalized regression model.}
\usage{
\method{fit_model}{easy_glmnet}(object)
}
\arguments{
\item{object}{A list of class \code{easy_glmnet}.}
}
\value{
A list of class \code{easy_glmnet}.
}
\description{
This function wraps the procedure for fitting a 
glmnet model and makes it accessible 
to the easyml core framework.
}
