% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fit_model}
\alias{fit_model}
\title{Fit model.}
\usage{
fit_model(object)
}
\arguments{
\item{object}{A list of class \code{easy_*}, where * is the name of the algorithm.}
}
\value{
A list of class \code{easy_*}, where * is the name of the algorithm.
}
\description{
The generic function  for fitting a model within 
the easyml core framework. Users can create their own 
fit_model function for their own class of model.
}
