% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_coefficients}
\alias{generate_coefficients}
\title{Generate coefficients.}
\usage{
generate_coefficients(object)
}
\arguments{
\item{object}{A list of class \code{easy_*}, where * is the name of the algorithm.}
}
\value{
A data.frame, the generated penalized regression model coefficients.
}
\description{
Generate coefficients.
}
\seealso{
Other generate: \code{\link{generate_model_performance}},
  \code{\link{generate_predictions}},
  \code{\link{generate_variable_importances}}
}
