% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_variable_importances}
\alias{generate_variable_importances}
\title{Generate variable importances for a model (if applicable).}
\usage{
generate_variable_importances(object)
}
\arguments{
\item{object}{A list of class \code{easy_*}, where * is the name of the algorithm.}
}
\value{
A data.frame, the generated variable importance scores.
}
\description{
Generate variable importances for a model (if applicable).
}
\seealso{
Other generate: \code{\link{generate_coefficients}},
  \code{\link{generate_model_performance}},
  \code{\link{generate_predictions}}
}
