% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet_extract_coefficients}
\alias{glmnet_extract_coefficients}
\title{Extract coefficients from a penalized regression model.}
\usage{
glmnet_extract_coefficients(results)
}
\arguments{
\item{results}{The results of \code{\link{glmnet_fit_model_gaussian}} or \code{\link{glmnet_fit_model_binomial}}.}
}
\value{
A data.frame of replicated penalized regression coefficients.
}
\description{
Extract coefficients from a penalized regression model.
}

