% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coefficients_processed}
\alias{plot_coefficients_processed}
\title{Plot processed coefficients.}
\usage{
plot_coefficients_processed(coefficients_processed)
}
\arguments{
\item{coefficients_processed}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
Plot processed coefficients.
}
\seealso{
Other plot: \code{\link{plot_metrics_binomial_area_under_curve}},
  \code{\link{plot_metrics_gaussian_mean_squared_error}},
  \code{\link{plot_metrics_gaussian_r2_score}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}}
}

