% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coefficients_processed}
\alias{plot_coefficients_processed}
\title{Plot penalized regression coefficients.}
\usage{
plot_coefficients_processed(coefficients_processed)
}
\arguments{
\item{coefficients_processed}{A data.frame, the output of the function \code{\link{process_coefficients}}.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
When calling \code{\link{easy_glmnet}}, coefficients from the 
\code{\link{generate_coefficients}} output are processed by the 
\code{\link{process_coefficients}}  function and generated into 
a plot. This plot tells us the direction, magnitude, and statistical
significance of each coefficient. Be careful using this plotting
method with datasets containing more than 20 variables as the plot 
may not render as nicely.
}
\seealso{
Other plot: \code{\link{plot_model_performance_binomial_auc_score}},
  \code{\link{plot_model_performance_gaussian_correlation_score}},
  \code{\link{plot_model_performance_gaussian_mse_score}},
  \code{\link{plot_model_performance_gaussian_r2_score}},
  \code{\link{plot_model_performance_histogram}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_roc_curve}},
  \code{\link{plot_variable_importances_processed}}
}
