% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metrics_binomial_area_under_curve}
\alias{plot_metrics_binomial_area_under_curve}
\title{TO BE EDITED.}
\usage{
plot_metrics_binomial_area_under_curve(aucs)
}
\arguments{
\item{aucs}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

