% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metrics_gaussian_mean_squared_error}
\alias{plot_metrics_gaussian_mean_squared_error}
\title{Plot mean squared error metrics.}
\usage{
plot_metrics_gaussian_mean_squared_error(mses)
}
\arguments{
\item{mses}{A vector, the mean squared error metrics to be plotted as a histogram.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
Plot mean squared error metrics.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_metrics_binomial_area_under_curve}},
  \code{\link{plot_metrics_gaussian_cor_score}},
  \code{\link{plot_metrics_gaussian_r2_score}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_variable_importances_processed}}
}

