% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metrics_gaussian_r2_score}
\alias{plot_metrics_gaussian_r2_score}
\title{TO BE EDITED.}
\usage{
plot_metrics_gaussian_r2_score(r2_scores)
}
\arguments{
\item{r2_scores}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

