% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_predictions_binomial}
\alias{plot_predictions_binomial}
\title{Plot binomial predictions.}
\usage{
plot_predictions_binomial(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
Plots a logistic plot of the ground truth (correct) target values 
and the estimated target values.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_model_performance_binomial_auc_score}},
  \code{\link{plot_model_performance_gaussian_correlation_score}},
  \code{\link{plot_model_performance_gaussian_mse_score}},
  \code{\link{plot_model_performance_gaussian_r2_score}},
  \code{\link{plot_model_performance_histogram}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_roc_curve}},
  \code{\link{plot_variable_importances_processed}}
}
