% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot ROC Curve.}
\usage{
plot_roc_curve(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values.}

\item{y_pred}{Estimated target values.}
}
\value{
A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.
}
\description{
Plot ROC Curve.
}
\seealso{
Other plot: \code{\link{plot_coefficients_processed}},
  \code{\link{plot_metrics_binomial_auc_score}},
  \code{\link{plot_metrics_gaussian_correlation_score}},
  \code{\link{plot_metrics_gaussian_mse_score}},
  \code{\link{plot_metrics_gaussian_r2_score}},
  \code{\link{plot_metrics_histogram}},
  \code{\link{plot_predictions_binomial}},
  \code{\link{plot_predictions_gaussian}},
  \code{\link{plot_variable_importances_processed}}
}

