% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinternet.R
\name{predict_model.easy_glinternet}
\alias{predict_model.easy_glinternet}
\title{Predict values for a penalized regression model with interactions.}
\usage{
\method{predict_model}{easy_glinternet}(object, newx = NULL)
}
\arguments{
\item{object}{A list of class \code{easy_glinternet}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values using the new data.
}
\description{
This function wraps the procedure for predicting values from 
a glinternet model and makes it accessible 
to the easyml core framework.
}
