% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{predict_model.easy_glmnet}
\alias{predict_model.easy_glmnet}
\title{Predict values for a penalized regression model.}
\usage{
\method{predict_model}{easy_glmnet}(object, newx = NULL)
}
\arguments{
\item{object}{A list of class \code{easy_glmnet}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values using the new data.
}
\description{
This function wraps the procedure for predicting values from 
a glmnet model and makes it accessible 
to the easyml core framework.
}
