% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_identity}
\alias{preprocess_identity}
\title{Preprocess data by leaving it exactly the way it is.}
\usage{
preprocess_identity(.data, categorical_variables = NULL)
}
\arguments{
\item{.data}{A data.frame; the data to be analyzed.}

\item{categorical_variables}{A logical vector; each value TRUE indicates that column in the data.frame is a categorical variable. Defaults to NULL.}
}
\value{
A list, containing one or two data.frames.
}
\description{
This function is the same as the identity function. Anything passed
to it is returned untouched.
}
\seealso{
Other preprocess: \code{\link{preprocess_scale}}
}
