% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_scale}
\alias{preprocess_scale}
\title{Preprocess data by scaling it.}
\usage{
preprocess_scale(.data, categorical_variables = NULL)
}
\arguments{
\item{.data}{A data.frame; the data to be analyzed.}

\item{categorical_variables}{A logical vector; each value TRUE indicates that column in the data.frame is a categorical variable. Defaults to NULL.}
}
\value{
A list, containing one or two data.frames.
}
\description{
This function takes either a data.frame or a list of data.frames. In 
the event of the first, this function takes the dataset and will
scale each column that is not categorical such that that column
has zero mean and unit variance. In the event of the second, this 
function takes the training dataset and will identify the parameters 
needed to scale the training dataset such that each column that 
is not categorical has zero mean and unit variance, and then will 
apply those parameters to each column in the testing dataset 
that is not categorical.
}
\seealso{
Other preprocess: \code{\link{preprocess_identity}}
}
