% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_coefficients}
\alias{process_coefficients}
\title{TO BE EDITED.}
\usage{
process_coefficients(coefs, n_samples, column_names,
  survival_rate_cutoff = 0.05)
}
\arguments{
\item{coefs}{TO BE EDITED.}

\item{n_samples}{TO BE EDITED.}

\item{survival_rate_cutoff}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

