% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_coefficients}
\alias{process_coefficients}
\title{Process coefficients.}
\usage{
process_coefficients(coefficients, survival_rate_cutoff = 0.05)
}
\arguments{
\item{coefficients}{A data.frame, the replicated coefficients.}

\item{survival_rate_cutoff}{A numeric vector of length one; for \code{\link{easy_glmnet}}, specifies the minimal threshold (as a percentage) a coefficient must appear out of n_samples. Defaults to 0.05.}
}
\value{
A data.frame; the replicated coefficients processed for easy plotting.
}
\description{
Takes the coefficients returned by the generate_coefficients function 
and prepares the coefficients for plotting.
}
\seealso{
Other utils: \code{\link{correlation_test}},
  \code{\link{process_variable_importances}},
  \code{\link{reduce_cores}},
  \code{\link{remove_variables}}
}
