% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{Prostate data.}
\format{Data frame with columns
\describe{
\item{lcavol}{To Be Edited}
\item{lweight}{To Be Edited}
\item{age}{To Be Edited}
\item{lbph}{To Be Edited}
\item{svi}{To Be Edited}
\item{lcp}{To Be Edited}
\item{gleason}{To Be Edited}
\item{pgg45}{To Be Edited}
\item{lpsa}{To Be Edited}
}}
\source{
https://www.tobeedited.com/
}
\usage{
prostate
}
\description{
Prostate data.
}
\examples{
  prostate
}
\keyword{datasets}

