% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest_predict_model}
\alias{random_forest_predict_model}
\title{Predict values for a random forest regression model.}
\usage{
random_forest_predict_model(results, newx = NULL)
}
\arguments{
\item{results}{The results of \code{\link{random_forest_fit_model_gaussian}} or \code{\link{random_forest_fit_model_binomial}}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values for a random_forest regression model using the new data.
}
\description{
Predict values for a random forest regression model.
}

