% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_coefficients}
\alias{replicate_coefficients}
\title{TO BE EDITED.}
\usage{
replicate_coefficients(fit_model, extract_coefficients, preprocess, X, y,
  categorical_variables = NULL, n_samples = 1000, progress_bar = TRUE,
  n_core = 1, ...)
}
\arguments{
\item{fit_model}{TO BE EDITED.}

\item{extract_coefficients}{TO BE EDITED.}

\item{X}{TO BE EDITED.}

\item{y}{TO BE EDITED.}

\item{n_samples}{TO BE EDITED.}

\item{progress_bar}{TO BE EDITED.}

\item{parallel}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

