% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_predictions}
\alias{replicate_predictions}
\title{Replicate predictions.}
\usage{
replicate_predictions(object)
}
\arguments{
\item{object}{TO BE EDITED.}
}
\value{
A list of matrixes, the replicated predictions.
}
\description{
Replicate predictions.
}
\seealso{
Other replicate: \code{\link{replicate_coefficients}},
  \code{\link{replicate_metrics}},
  \code{\link{replicate_variable_importances}}
}

