% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_predictions}
\alias{replicate_predictions}
\title{Replicate predictions.}
\usage{
replicate_predictions(fit_model, predict_model, preprocess, X_train, y_train,
  X_test, categorical_variables = NULL, n_samples = 1000,
  progress_bar = TRUE, n_core = 1, ...)
}
\arguments{
\item{fit_model}{A function; the function for fitting a model to the data.}

\item{predict_model}{A function; the function for generating predictions from a fitted model.}

\item{preprocess}{A function; the function for preprocessing the data. Defaults to NULL.}

\item{X_train}{A matrix; the independent variables sampled to a training set.}

\item{y_train}{A vector; the dependent variable sampled to a training set.}

\item{X_test}{A matrix; the independent variables sampled to a testing set.}

\item{categorical_variables}{A logical vector; each value TRUE indicates that column in the data.frame is a categorical variable. Defaults to NULL.}

\item{n_samples}{An integer vector of length one; specifies the number of times the coefficients and predictions should be replicated. Defaults to 1000L.}

\item{progress_bar}{A logical vector of length one; specifies whether to display a progress bar during calculations. Defaults to TRUE.}

\item{n_core}{An integer vector of length one; specifies the number of cores to use for this analysis. Currenly only works on Mac OSx and Unix/Linux systems. Defaults to 1L.}

\item{...}{The arguments to be passed to the algorithm specified.}
}
\value{
A list of matrixes, the replicated predictions.
}
\description{
Replicate predictions.
}
\seealso{
Other replicate: \code{\link{replicate_coefficients}},
  \code{\link{replicate_metrics}},
  \code{\link{replicate_variable_importances}}
}

