% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_variable_importances}
\alias{replicate_variable_importances}
\title{Replicate variable importances.}
\usage{
replicate_variable_importances(fit_model, preprocess, X, y,
  categorical_variables = NULL, ...)
}
\arguments{
\item{fit_model}{TO BE EDITED.}

\item{preprocess}{A function; the function for preprocessing the data. Defaults to NULL.}

\item{X}{A matrix; the independent variables.}

\item{y}{A vector; the dependent variable.}

\item{categorical_variables}{A logical vector; each value TRUE indicates that column in the data.frame is a categorical variable. Defaults to NULL.}

\item{...}{The arguments to be passed to the algorithm specified.}
}
\value{
TO BE EDITED.
}
\description{
Replicate variable importances.
}
\seealso{
Other replicate: \code{\link{replicate_coefficients}},
  \code{\link{replicate_metrics}},
  \code{\link{replicate_predictions}}
}

