% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_categorical_variables}
\alias{set_categorical_variables}
\title{Set categorical variables.}
\usage{
set_categorical_variables(column_names, categorical_variables = NULL)
}
\arguments{
\item{column_names}{A character vector; the column names of the data for this analysis.}

\item{categorical_variables}{A character vector; the variables that are categorical. Defaults to NULL.}
}
\value{
TO BE EDITED.
}
\description{
Set categorical variables.
}

