% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_categorical_variables}
\alias{set_categorical_variables}
\title{Set categorical variables.}
\usage{
set_categorical_variables(column_names, categorical_variables = NULL)
}
\arguments{
\item{column_names}{A character vector; the column names of the data for this analysis.}

\item{categorical_variables}{A character vector; the variables that are categorical. Defaults to NULL.}
}
\value{
NULL, or if \code{categorical_variables} is not NULL, then a logical vector of length \code{length(column_names} where TRUE represents that column is a categorical variable.
}
\description{
This helper functions determines a logical boolean vector based on the column names 
and the designation for which ones are categorical variables.
}
\seealso{
Other setters: \code{\link{set_column_names}},
  \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}}, \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}}
}
