% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_coefficients_boolean}
\alias{set_coefficients_boolean}
\title{Set coefficients boolean.}
\usage{
set_coefficients_boolean(algorithm)
}
\arguments{
\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}
}
\value{
A logical vector of length one; whether coefficients should be replicated for this analyis or not.
}
\description{
Set coefficients boolean.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_extract_coefficients}},
  \code{\link{set_extract_variable_importances}},
  \code{\link{set_fit_model}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}},
  \code{\link{set_metrics_boolean}},
  \code{\link{set_parallel}},
  \code{\link{set_plot_metrics}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_predict_model}},
  \code{\link{set_predictions_boolean}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}},
  \code{\link{set_variable_importances_boolean}}
}

