% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_column_names}
\alias{set_column_names}
\title{TO BE EDITED.}
\usage{
set_column_names(column_names, dependent_variable, exclude_variables = NULL,
  preprocessor = NULL, categorical_variables = NULL)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

