% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_column_names}
\alias{set_column_names}
\title{Set column names.}
\usage{
set_column_names(column_names, dependent_variable, preprocess = NULL,
  exclude_variables = NULL, categorical_variables = NULL)
}
\arguments{
\item{column_names}{A character vector; the column names of the data for this analysis.}

\item{dependent_variable}{A character vector of length one; the dependent variable for this analysis.}

\item{preprocess}{A function; the function for preprocessing the data. Defaults to NULL.}

\item{exclude_variables}{A character vector; the variables from the data set to exclude. Defaults to NULL.}

\item{categorical_variables}{A character vector; the variables that are categorical. Defaults to NULL.}
}
\value{
TO BE EDITED.
}
\description{
Set column names.
}

