% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_column_names}
\alias{set_column_names}
\title{Set column names.}
\usage{
set_column_names(column_names, dependent_variable, preprocess = NULL,
  exclude_variables = NULL, categorical_variables = NULL)
}
\arguments{
\item{column_names}{A character vector; the column names of the data for this analysis.}

\item{dependent_variable}{A character vector of length one; the dependent variable for this analysis.}

\item{preprocess}{A function; the function for preprocessing the data. Defaults to NULL.}

\item{exclude_variables}{A character vector; the variables from the data set to exclude. Defaults to NULL.}

\item{categorical_variables}{A character vector; the variables that are categorical. Defaults to NULL.}
}
\value{
The updated columns, in the correct order for preprocessing.
}
\description{
Set column names.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_coefficients_boolean}},
  \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_extract_coefficients}},
  \code{\link{set_extract_variable_importances}},
  \code{\link{set_fit_model}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}},
  \code{\link{set_metrics_boolean}},
  \code{\link{set_parallel}},
  \code{\link{set_plot_metrics}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_predict_model}},
  \code{\link{set_predictions_boolean}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}},
  \code{\link{set_variable_importances_boolean}}
}

