% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_cores}
\alias{set_cores}
\title{Set cores.}
\usage{
set_cores(n_core)
}
\arguments{
\item{n_core}{An integer vector of length one; specifies the number of cores to use for this analysis. Currently only works on Mac OSx and Unix/Linux systems. Defaults to 1L.}
}
\value{
NULL.
}
\description{
Please note this affects global state and sets the number of cores by running \code{options(mc.cores = n_core)}.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}}, \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}}
}
