% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_extract_coefficients}
\alias{set_extract_coefficients}
\title{Set extract coefficients function.}
\usage{
set_extract_coefficients(algorithm, family)
}
\arguments{
\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}

\item{family}{A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".}
}
\value{
A function; the function for extracting coefficients from a model.
}
\description{
Sets the function responsible for extracting coefficients from a model.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_coefficients_boolean}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_extract_variable_importances}},
  \code{\link{set_fit_model}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}},
  \code{\link{set_metrics_boolean}},
  \code{\link{set_parallel}},
  \code{\link{set_plot_metrics}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_predict_model}},
  \code{\link{set_predictions_boolean}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}},
  \code{\link{set_variable_importances_boolean}}
}

