% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_independent_variables}
\alias{set_independent_variables}
\title{Set independent variables.}
\usage{
set_independent_variables(.data, dependent_variable)
}
\arguments{
\item{.data}{A data.frame; the data to be analyzed.}

\item{dependent_variable}{A character vector of length one; the dependent variable for this analysis.}
}
\value{
A data.frame, the independent variables of the analysis.
}
\description{
This helper functions isolates the independent variables in a data.frame.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}}, \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}}
}
