% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_looper_}
\alias{set_looper_}
\title{Set looper.}
\usage{
set_looper_(progress_bar = FALSE, parallel = FALSE)
}
\arguments{
\item{progress_bar}{A logical vector of length one; specifies whether to display a progress bar during calculations. Defaults to FALSE.}

\item{parallel}{A logical vector of length one; specifies whether to run calculations in parallel. Defaults to FALSE.}
}
\value{
The looper to use depending on progress bar and whether to run in parallel or not.
}
\description{
This function decides which looper (a functional like lapply) to run. This function does not affect global state.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper}}, \code{\link{set_measure}},
  \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}}
}
