% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_measure}
\alias{set_measure}
\title{Set measure function.}
\usage{
set_measure(measure = NULL, algorithm, family)
}
\arguments{
\item{measure}{A function; the function for measuring the results. Defaults to NULL.}

\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}

\item{family}{A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".}
}
\value{
A function; the function for measuring the results.
}
\description{
Sets the function responsible for measuring the results.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}}
}
