% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_plot_predictions}
\alias{set_plot_predictions}
\title{Set plot predictions function.}
\usage{
set_plot_predictions(algorithm, family)
}
\arguments{
\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}

\item{family}{A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".}
}
\value{
A function; the function for plotting the predictions generated from a fitted model.
}
\description{
Sets the function responsible for plotting the predictions generated from a fitted model.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}}, \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}}
}
