% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_preprocess}
\alias{set_preprocess}
\title{Set preprocess function.}
\usage{
set_preprocess(preprocess = NULL, algorithm)
}
\arguments{
\item{preprocess}{A function; the function for preprocessing the data. Defaults to NULL.}

\item{algorithm}{A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").}
}
\value{
A function; the function for preprocessing the data.
}
\description{
Sets the function responsible for preprocessing the data.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}}, \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_random_state}},
  \code{\link{set_resample}}
}
