% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_random_state}
\alias{set_random_state}
\title{Set random state.}
\usage{
set_random_state(random_state = NULL)
}
\arguments{
\item{random_state}{An integer vector of length one; specifies the seed to be used for the analysis. Defaults to NULL.}
}
\value{
NULL.
}
\description{
Sets the random state to a specific seed. Please note this function affects global state.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}}, \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_preprocess}}, \code{\link{set_resample}}
}
