% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_resample}
\alias{set_resample}
\title{Set resample function.}
\usage{
set_resample(resample = NULL, family = NULL)
}
\arguments{
\item{resample}{A function; the function for resampling the data. Defaults to NULL.}

\item{family}{A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".}
}
\value{
A function; the function for resampling the data.
}
\description{
Sets the function responsible for resampling the data.
}
\seealso{
Other setters: \code{\link{set_categorical_variables}},
  \code{\link{set_column_names}}, \code{\link{set_cores}},
  \code{\link{set_dependent_variable}},
  \code{\link{set_independent_variables}},
  \code{\link{set_looper_}}, \code{\link{set_looper}},
  \code{\link{set_measure}}, \code{\link{set_parallel}},
  \code{\link{set_plot_model_performance}},
  \code{\link{set_plot_predictions}},
  \code{\link{set_preprocess}},
  \code{\link{set_random_state}}
}
