% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{support_vector_machine_fit_model_gaussian}
\alias{support_vector_machine_fit_model_gaussian}
\title{Fit support vector machine gaussian regression model.}
\usage{
support_vector_machine_fit_model_gaussian(X, y, ...)
}
\arguments{
\item{X}{a data matrix, a vector, or a sparse matrix (object of class Matrix provided by the Matrix package, or of class matrix.csr provided by the SparseM package, or of class simple_triplet_matrix provided by the slam package).}

\item{y}{a response vector with one label for each row/component of x. Can be either a factor (for classification tasks) or a numeric vector (for regression).}

\item{...}{Arguments to be passed to \code{\link[e1071]{svm}}. See that function's documentation for more details.}
}
\value{
The fitted model.
}
\description{
Fit support vector machine gaussian regression model.
}

