% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{support_vector_machine_predict_model}
\alias{support_vector_machine_predict_model}
\title{Predict values for a support vector machine regression model.}
\usage{
support_vector_machine_predict_model(results, newx)
}
\arguments{
\item{results}{TO BE EDITED.}

\item{newx}{TO BE EDITED.}
}
\value{
TO BE EDITED.
}
\description{
Predict values for a support vector machine regression model.
}

