% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{support_vector_machine_predict_model}
\alias{support_vector_machine_predict_model}
\title{Predict values for a support vector machine regression model.}
\usage{
support_vector_machine_predict_model(results, newx = NULL)
}
\arguments{
\item{results}{The results of \code{\link{support_vector_machine_fit_model_gaussian}} or \code{\link{support_vector_machine_fit_model_binomial}}.}

\item{newx}{A data.frame, the new data to use for predictions.}
}
\value{
A vector, the predicted values for a support_vector_machine regression model using the new data.
}
\description{
Predict values for a support vector machine regression model.
}

